+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.577] - 2026-04-26
+
+- [Make it possible to select several output destinations](https://github.com/BetterThanTomorrow/calva/issues/2819)
+- [Support file paths as output destinations](https://github.com/BetterThanTomorrow/calva/issues/3185)
+
 ## [2.0.576] - 2026-04-25
 
 - [Make loadFile command accept an argument for which file to load](https://github.com/BetterThanTomorrow/calva/issues/3182)